+++ src/libs/USBDevice/USBMSD/USBMSD.cpp
@@ -281,15 +281,15 @@ bool USBMSD::inquiryRequest (void) {
 
 bool USBMSD::readFormatCapacity() {
     uint8_t capacity[] = { 0x00, 0x00, 0x00, 0x08,
-                           (BlockCount >> 24) & 0xff,
-                           (BlockCount >> 16) & 0xff,
-                           (BlockCount >> 8) & 0xff,
-                           (BlockCount >> 0) & 0xff,
+                            ((uint8_t) ((BlockCount >> 24) & 0xff)),
+                            ((uint8_t) ((BlockCount >> 16) & 0xff)),
+                            ((uint8_t) ((BlockCount >> 8) & 0xff)),
+                            ((uint8_t) ((BlockCount >> 0) & 0xff)),
 
-                           0x02,
-                           (BlockSize >> 16) & 0xff,
-                           (BlockSize >> 8) & 0xff,
-                           (BlockSize >> 0) & 0xff,
+                            0x02,
+                            ((uint8_t) ((BlockSize >> 16) & 0xff)),
+                            ((uint8_t) ((BlockSize >> 8) & 0xff)),
+                            ((uint8_t) ((BlockSize >> 0) & 0xff)),
                          };
     if (!write(capacity, sizeof(capacity))) {
         return false;
@@ -300,15 +300,15 @@ bool USBMSD::readFormatCapacity() {
 
 bool USBMSD::readCapacity (void) {
     uint8_t capacity[] = {
-        ((BlockCount - 1) >> 24) & 0xff,
-        ((BlockCount - 1) >> 16) & 0xff,
-        ((BlockCount - 1) >> 8) & 0xff,
-        ((BlockCount - 1) >> 0) & 0xff,
+        ((uint8_t) (((BlockCount - 1) >> 24) & 0xff)),
+        ((uint8_t) (((BlockCount - 1) >> 16) & 0xff)),
+        ((uint8_t) (((BlockCount - 1) >> 8) & 0xff)),
+        ((uint8_t) (((BlockCount - 1) >> 0) & 0xff)),
 
-        (BlockSize >> 24) & 0xff,
-        (BlockSize >> 16) & 0xff,
-        (BlockSize >> 8) & 0xff,
-        (BlockSize >> 0) & 0xff,
+        ((uint8_t) ((BlockSize >> 24) & 0xff)),
+        ((uint8_t) ((BlockSize >> 16) & 0xff)),
+        ((uint8_t) ((BlockSize >> 8) & 0xff)),
+        ((uint8_t) ((BlockSize >> 0) & 0xff)),
     };
     if (!write(capacity, sizeof(capacity))) {
         return false;
